backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
SRTrain
python-dotenv
httpx
keyring
